Update frontend.py to adjust how play/pause images are shown on the Pirate Audio display
Prior to this change, the image displayed on the screen next to the "A" button seemed opposite of what one would expect. When music is playing it showed the ">" (play) image and when the music is paused it showed the "||" (pause) image. One would expect that image to show the user what they'd like the system to do when they PRESS the button (NOT what the system is currently doing). This change makes it consistent with how other media players behave in this regard.